+++ src/main/java/webserver/RequestHandler.java
@@ -6,6 +6,7 @@ import java.net.Socket;
 import com.google.common.base.Charsets;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
+import webserver.request.domain.RequestLine;
 
 public class RequestHandler implements Runnable {
     private static final Logger logger = LoggerFactory.getLogger(RequestHandler.class);
@@ -23,6 +24,14 @@ public class RequestHandler implements Runnable {
         try (InputStream in = connection.getInputStream(); OutputStream out = connection.getOutputStream()) {
             // TODO 사용자 요청에 대한 처리는 이 곳에 구현하면 된다.
             showRequest(in);
+
+//            BufferedReader br = new BufferedReader(new InputStreamReader(in, Charsets.UTF_8));
+//            String line = br.readLine();
+//            logger.debug("request line : {}", line);
+//            while(!line.equals("")) {
+//                line = br.readLine();
+//                logger.debug("header : {}", line);
+//            }
             
             DataOutputStream dos = new DataOutputStream(out);
             byte[] body = "Hello World".getBytes();
@@ -37,13 +46,16 @@ public class RequestHandler implements Runnable {
         try {
             BufferedReader br = new BufferedReader(new InputStreamReader(is, Charsets.UTF_8));
             String line = br.readLine();
+            RequestLine requestLine = RequestLine.from(line);
+            logger.debug(requestLine.toString());
 
             StringBuilder stringBuilder = new StringBuilder();
-            do{
-                stringBuilder.append(line + "\n");
+            while(!line.equals("")) {
                 line = br.readLine();
-            } while(line != null);
-            System.out.println(stringBuilder);
+                stringBuilder.append(line + "\n");
+            }
+
+            logger.debug(String.valueOf(stringBuilder));
         } catch (IOException e) {
             logger.error(e.getMessage());
         }
