+++ src/main/java/webserver/request/domain/HttpMethod.java
@@ -1,4 +1,4 @@
-package webserver.domain;
+package webserver.request.domain;
 
 public enum HttpMethod {
     GET,
