+++ src/main/java/webserver/request/domain/HttpRequest.java
@@ -1,0 +1,4 @@
+package webserver.request.domain;
+
+public class HttpRequest {
+}
