+++ src/test/java/webserver/HttpRequestTest.java
@@ -14,4 +14,5 @@ public class HttpRequestTest {
         ResponseEntity<String> response = restTemplate.getForEntity("http://localhost:8080", String.class);
         assertThat(response.getStatusCode()).isEqualTo(HttpStatus.OK);
     }
+
 }
